lua: minor refactoring of the ocl-lua package
* added header files for classes LuaComponent and LuaService (with and without tlsf)
* added member function getLuaState() and class LuaStateHandle to access
  the Lua state from outside with proper locking
* split rttlua-* executable implementation into a separate source file rttlua.cpp
  The binaries link to the respective component library.


Conflicts:
	lua/CMakeLists.txt
	lua/plugins/CMakeLists.txt
