+++ lua/CMakeLists.txt
@@ -1,9 +1,9 @@
 
 option(BUILD_LUA_RTT "Build Lua RTT bindings" ON)
-cmake_dependent_option(BUILD_LUA_RTT_DYNAMIC_MODULES "Build dynamic loadable Lua plugins for rtt and deployer (experimental!)" OFF "BUILD_LUA_RTT" OFF)
+cmake_dependent_option(BUILD_LUA_RTT_DYNAMIC_MODULES "Build dynamic loadable Lua plugins for rtt and deployer" OFF "BUILD_LUA_RTT" OFF)
 
-cmake_dependent_option(BUILD_LUA_RTT_TLSF "Build TLSF enabled Lua version (experimental!)" ON "BUILD_LUA_RTT" OFF)
-cmake_dependent_option(BUILD_LUA_RTT_CORBA "Build CORBA enabled Lua version (experimental!)" ON "BUILD_LUA_RTT" OFF)
+cmake_dependent_option(BUILD_LUA_RTT_TLSF "Build TLSF enabled Lua version" ON "BUILD_LUA_RTT" OFF)
+cmake_dependent_option(BUILD_LUA_RTT_CORBA "Build CORBA enabled Lua version" ON "BUILD_LUA_RTT" OFF)
 cmake_dependent_option(LUA_RTT_TYPEINFO_CACHING "Let Lua RTT bindings do Typeinfo Caching" ON "BUILD_LUA_RTT" OFF)
 
 if(BUILD_LUA_RTT)
@@ -43,12 +43,12 @@ if(BUILD_LUA_RTT)
       add_definitions("-DTYPEINFO_CACHING")
     endif(LUA_RTT_TYPEINFO_CACHING)
 
-    orocos_component(orocos-ocl-lua rtt.cpp LuaComponent.cpp )
+    orocos_component(orocos-ocl-lua rtt.cpp LuaComponent.cpp LuaStateHandle.cpp )
     set_target_properties(orocos-ocl-lua PROPERTIES SOVERSION ${OCL_SOVERSION})
-    orocos_executable(rttlua rtt.cpp LuaComponent.cpp)
-    target_link_libraries(rttlua lua-repl orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES} ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
     target_link_libraries(orocos-ocl-lua ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES} )
-    set_target_properties(orocos-ocl-lua PROPERTIES COMPILE_FLAGS "-DOCL_COMPONENT_ONLY")
+
+    orocos_executable(rttlua rttlua.cpp)
+    target_link_libraries(rttlua lua-repl orocos-ocl-lua orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES} ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
     install(TARGETS rttlua RUNTIME DESTINATION bin)
 
     configure_file(../bin/rttlua.in ../bin/rttlua @ONLY)
@@ -70,8 +70,8 @@ if(BUILD_LUA_RTT)
     # CORBA version
     if(BUILD_LUA_RTT_CORBA AND OROCOS-RTT_CORBA_FOUND)
 
-      orocos_executable(rttlua-corba rtt.cpp LuaComponent.cpp)
-      target_link_libraries(rttlua-corba lua-repl orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_CORBA_LIBRARIES} orocos-ocl-deployment-corba ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
+      orocos_executable(rttlua-corba rttlua.cpp)
+      target_link_libraries(rttlua-corba orocos-ocl-lua lua-repl orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_CORBA_LIBRARIES} orocos-ocl-deployment-corba ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
       set_target_properties(rttlua-corba PROPERTIES COMPILE_FLAGS "-DLUA_RTT_CORBA")
       install(TARGETS rttlua-corba RUNTIME DESTINATION bin)
 
@@ -83,22 +83,21 @@ if(BUILD_LUA_RTT)
     # TLSF version
     if(BUILD_LUA_RTT_TLSF)
 
-      orocos_component(orocos-ocl-lua-tlsf rtt.cpp LuaComponent.cpp )
+      orocos_component(orocos-ocl-lua-tlsf rtt.cpp LuaComponent.cpp LuaStateHandle.cpp )
       set_target_properties(orocos-ocl-lua-tlsf PROPERTIES SOVERSION ${OCL_SOVERSION})
-      orocos_executable(rttlua-tlsf rtt.cpp LuaComponent.cpp)
+      target_link_libraries(orocos-ocl-lua-tlsf tlsf_rtt ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES})
+      set_target_properties(orocos-ocl-lua-tlsf PROPERTIES COMPILE_FLAGS "-DLUA_RTT_TLSF")
+
       add_library(lua-repl-tlsf STATIC lua-repl.c)
-      target_link_libraries(lua-repl-tlsf ${READLINE_LIBRARY} ${CURSES_LIBRARY} "dl" ${HISTORY_LIBRARY} ${LUA_LIBRARIES})
+      target_link_libraries(lua-repl-tlsf orocos-ocl-lua-tlsf ${READLINE_LIBRARY} ${CURSES_LIBRARY} "dl" ${HISTORY_LIBRARY} ${LUA_LIBRARIES})
       set_target_properties(lua-repl-tlsf PROPERTIES COMPILE_FLAGS "-DOROCOS_TARGET=${OROCOS_TARGET} -DLUA_RTT_TLSF ${LUA_REPL_FLAGS}")
 
       add_library(tlsf_rtt STATIC tlsf.c tlsf_rtt.c)
       set_target_properties(tlsf_rtt PROPERTIES COMPILE_FLAGS "-DTLSF_STATISTIC -DCHECK_DOUBLE_FREE -fPIC -Wextra -Wall -Werror")
 
-      set_target_properties(orocos-ocl-lua-tlsf PROPERTIES COMPILE_FLAGS "-DOCL_COMPONENT_ONLY -DLUA_RTT_TLSF")
+      orocos_executable(rttlua-tlsf rttlua.cpp)
+      target_link_libraries(rttlua-tlsf orocos-ocl-lua-tlsf tlsf_rtt lua-repl-tlsf orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES} ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
       set_target_properties(rttlua-tlsf PROPERTIES COMPILE_FLAGS "-DLUA_RTT_TLSF")
-
-      target_link_libraries(rttlua-tlsf tlsf_rtt lua-repl-tlsf orocos-ocl-deployment ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES} ${EXTRA_DEPS} ${EXTRA_LIBRARIES} )
-      target_link_libraries(orocos-ocl-lua-tlsf tlsf_rtt ${LUA_LIBRARIES} ${OROCOS-RTT_LIBRARIES})
-
       install(TARGETS rttlua-tlsf RUNTIME DESTINATION bin)
       install(FILES modules/tlsf_ext.lua DESTINATION share/lua/5.1/)
 
@@ -106,7 +105,19 @@ if(BUILD_LUA_RTT)
       install(PROGRAMS ${CMAKE_CURRENT_BINARY_DIR}/../bin/rttlua-tlsf DESTINATION bin) # The shell script
     endif(BUILD_LUA_RTT_TLSF)
 
-    add_subdirectory( plugins )
+    orocos_plugin( rttlua-plugin LuaService.cpp LuaStateHandle.cpp rtt.cpp )
+    set_target_properties(rttlua-plugin PROPERTIES SOVERSION ${OCL_SOVERSION})
+    target_link_libraries(rttlua-plugin ${LUA_LIBRARIES})
+
+    # TLSF version
+    if(BUILD_LUA_RTT_TLSF)
+      orocos_plugin( rttlua-tlsf-plugin LuaService.cpp LuaStateHandle.cpp rtt.cpp )
+      set_target_properties(rttlua-tlsf-plugin PROPERTIES SOVERSION ${OCL_SOVERSION})
+      target_link_libraries(rttlua-tlsf-plugin tlsf_rtt ${LUA_LIBRARIES})
+      set_target_properties(rttlua-tlsf-plugin PROPERTIES COMPILE_FLAGS "-DLUA_RTT_TLSF")
+    endif(BUILD_LUA_RTT_TLSF)
+
+    orocos_install_headers( LuaComponent.hpp LuaService.hpp LuaStateHandle.hpp )
     add_subdirectory( testing )
 
     orocos_generate_package()
